requirements.txt
requmatplotlib==3.3.4
mlxtend==0.19.0
numpy==1.19.2
pandas==1.3.4
pwlf==2.0.5
PyYAML==6.0
scikit_learn==1.2.0
scipy~=1.5.4
sympy==1.9
torch==1.7.1
scikit-learn==0.24.2
openturns==1.19

